backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
packaging>=24.1
function_schema>=0.4.5
celery==5.4.0
python-decouple==3.8
redis==5.0.7
qdrant-client==1.10.1
google-generativeai>=0.8.0
openai>=1.56.1
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
tavily-python==0.5.0
python-dotenv
pymongo
nltk
elasticsearch